+++ data/src/main/java/com/moneymong/moneymong/data/datasource/mock/AgencyRemoteDataSourceMock.kt
@@ -24,7 +24,7 @@ class AgencyRemoteDataSourceMock @Inject constructor(
     }
 
     override suspend fun agencyCodeNumbers(
-        agencyId: Int,
+        agencyId: Long,
         codeNumbers: AgencyJoinRequest
     ): Result<AgencyJoinResponse> {
         return agencyJoinApi.agencyCodeNumbers("Bearer eyJhbGciOiJIUzI1NiJ9.eyJyb2xlIjoiUk9MRV9VU0VSIiwidXNlcklkIjozLCJpYXQiOjE3MDQ3MTU0NTEsImV4cCI6MTczNjI3MzA1MX0.2yYEy71Gz4YIz0DYzlx0glYMgZA0JAZs05jsVRvvQx4", agencyId, codeNumbers)
